### 2018-09-06 #930 @whedon
[PRE REVIEW]: qtools: A Python toolset for the Q molecular simulation package.
**Submitting author:** @mpurg (<a href="http://orcid.org/0000-0003-4647-6103">Miha Purg</a>)
**Repository:** <a href="https://github.com/mpurg/qtools" target ="_blank">https://github.com/mpurg/qtools</a>
**Version:** v0.6.1
