### 2018-09-06 #930 @mpurg
Hi @kyleniemeyer,

Based on declared topics of research, I can suggest the following users
from the list of joss reviewers:
@richardjgowers
@fdroessler
@malramsay64

Additionally, I can suggest an expert in the topic specific to the
manuscript:
@esguerra

Best regards,
Miha

V V čet., 6. sep. 2018 ob 18:19 je oseba Arfon Smith <
notifications@github.com> napisala:

> 👋 @kyleniemeyer <https://github.com/kyleniemeyer> - the submitting
> author suggest you as the handling editor.
>
