### 2018-09-10 #930 @esguerra
Hi @kyleniemeyer and @mpurg. I would be glad to help or help the reviewers in whichever way @kyleniemeyer  seems fit. For full disclosure I know the author of the paper and we work in the same institution, but in different groups and departments.
