### 2018-09-11 #935 @kyleniemeyer
@mpurg @malramsay64 @esguerra @fdroessler 👋  you each have individual checklists above. Detailed review criteria are at https://joss.readthedocs.io/en/latest/review_criteria.html
