### 2018-09-11 c4728e5
Describe link to Travis CI
Ref #8
