### 2018-09-11 e118a8c
Describe link to manuscript
Including a mention that the article is paywalled.

Ref #8
