### 2018-09-11 0473b6a
Put badges below title
This is the way most repositories on GitHub are doing this, with an example being [Numpy](https://github.com/numpy/numpy). Just makes for a more consistent experience.
