### 2018-09-11 #935 @malramsay64
@mpurg I have taken an initial look and raised a number of issues. To be clear, only the issues with '[review]' in the title are the ones which require action based on the review criteria. The others are more suggestions which I think would make the software better. I am happy to help fix or discuss any issues.
