### 2018-09-18 #935 @esguerra
@mpurg @malramsay64  @fdroessle @kyleniemeyer
Sorry  this reply has taken so long.
Only until now do I have time to read the COI policy in detail.
The author of this software and I are coauthors in a paper published less than four years ago, so,  I cannot act as a reviewer, sadly.

Thanks for the invitation and what a great idea this is,

Mauricio
