### 2018-09-24 #935 @malramsay64
@kyleniemeyer I have some time this week set aside to continue my review. I had somewhat put it on the side expecting some response from @mpurg to the issues I have raised.
