### 2018-09-24 70ed7d6
Modernize Qpyl code+tests, works with py2.7.6
