+++ tests/Qpyl/plotdata_test.py
@@ -2,6 +2,7 @@
 # py.test test functions
 #########################
 
+from __future__ import absolute_import
 import re
 import pytest
 

[…]
