### 2018-09-25 1e4784c
Qpyl tests clean with py3.7
Updated qscripts-cli and tests to work with 2.7.6

QMapper.details and QGroubContrib.details were finally cleaned of
CLI-dependent output.
