+++ tests/Qpyl/plotdata_test.py
@@ -6,12 +6,24 @@ from __future__ import absolute_import
 import re
 import pytest
 
-from Qpyl.plotdata import PlotData, PlotDataError, PlotDataJSONDecoder
+from Qpyl.plotdata import PlotData, PlotDataError
+from Qpyl.plotdata import PlotDataJSONEncoder, PlotDataJSONDecoder
 
 def is_close(a, b, rel_tol=1e-09, abs_tol=0.0):
     return abs(a-b) <= max(rel_tol * max(abs(a), abs(b)), abs_tol)
 
 class TestPlotData:
+    def test_encoder(self):
+        pl = PlotData("random_title", "bar", "xlab", "ylab", "zlab")
+        pl.add_subplot("1.subplot", [1,2,3,4], [2,4,8,16],
+                       [2,4,8,16], [1,2,3,4])
+        pl.add_subplot("2.subplot", [1,2,3,4], [3,4,5,6],
+                       [2,3,4,5], [1,1,1,1])
+        jsonenc = PlotDataJSONEncoder(indent=2, separators=(',', ': '))
+        ref_data = open("data/PlotData.json").read()
+        assert ref_data == jsonenc.encode(pl)
+
+
     def test_decoder(self):
         # decode json to PlotData
         qaf_data = open("data/qaf.PlotData.json").read()
