### 2018-09-26 ba333bc
Added all __future__ imports to Qpyl and checked the individual changes from modernize.
