### 2018-09-26 f3abb83
qscripts-cli: tests work with py27 after modernize
