+++ qscripts-cli/q_ffld2q.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
 # MIT License
@@ -37,6 +37,8 @@
 #   $SCHRODINGER/utilities/ffld_server -ipdb paraoxon.pdb -print_parameters -version 14 > paraoxon.ffld11
 #
 #
+from __future__ import absolute_import
+from __future__ import print_function
 from qscripts_config import __version__, QScriptsConfig as QScfg
 
 import sys
@@ -51,6 +53,8 @@ from Qpyl.core.qpotential import torsion_energy
 from Qpyl.core.qstructure import QStruct, QStructError
 from Qpyl.core.qtopology import QTopology, QTopologyError
 from Qpyl.common import backup_file, init_logger, get_version_full
+import six
+from six.moves import range
 
 logger = init_logger('Qpyl')
 
@@ -90,9 +94,9 @@ if len(sys.argv) == 1:
 
 args = parser.parse_args()
 
-for k, v in vars(args).iteritems():
+for k, v in six.iteritems(vars(args)):
     if k in ['ffld_output', 'pdb'] and not os.path.lexists(v):
-        print "FATAL! File '{}' doesn't exist.".format(v)
+        print("FATAL! File '{}' doesn't exist.".format(v))
         sys.exit(1)
 
 
@@ -106,26 +110,26 @@ qprm = QPrm("oplsaa", ignore_errors=args.ignore_errors)
 try:
     qstruct = QStruct(args.pdb, "pdb", ignore_errors=args.ignore_errors)
 except QStructError as err:
-    print "FATAL! Problem with pdb: {}".format(err)
+    print("FATAL! Problem with pdb: {}".format(err))
     sys.exit(1)
 
 try:
     qprm.read_ffld(args.ffld_output, qstruct)
 except QPrmError as err:
-    print "FATAL! Problem with ffld file: {}".format(err)
+    print("FATAL! Problem with ffld file: {}".format(err))
     sys.exit(1)
 
 
 try:
     qlib.read_ffld(args.ffld_output, qstruct)
 except QLibError as err:
-    print "FATAL! Problem with ffld file: {}".format(err)
+    print("FATAL! Problem with ffld file: {}".format(err))
     sys.exit(1)
 
 try:
     qtop = QTopology(qlib, qprm, qstruct)
 except QTopologyError as err:
-    print "FATAL! Problem building the topology: {}".format(err)
+    print("FATAL! Problem building the topology: {}".format(err))
     sys.exit(1)
 
 #
@@ -186,7 +190,7 @@ for p, de, v, v0 in all_prms:
 #
 # output some information
 #
-print """
+print("""
 Details about the system:
 
 Bonds: {n[bonds]}
@@ -199,7 +203,7 @@ Angle energy: total {t[angles]:.2f}, max {m[angles]:.2f}
 Torsion energy: total {t[torsions]:.2f}, max {m[torsions]:.2f}
 Improper energy: total {t[impropers]:.2f}, max {m[impropers]:.2f}
 
-""".format(n=nprm, t=total_e, m=max_e)
+""".format(n=nprm, t=total_e, m=max_e))
 
 #
 # write the files
@@ -208,10 +212,10 @@ libfn = args.output_basename + ".lib"
 prmfn = args.output_basename + ".prm"
 prmchkfn = args.output_basename + ".prm.chk"
 
-print "Writing the library file: {}".format(libfn)
+print("Writing the library file: {}".format(libfn))
 backup = backup_file(libfn)
 if backup:
-    print "# Backed up '{}' to '{}'".format(libfn, backup)
+    print("# Backed up '{}' to '{}'".format(libfn, backup))
[…]
