+++ qscripts-cli/q_genrelax.py
@@ -1,4 +1,4 @@
-#!/usr/bin/env python2
+#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 #
 # MIT License
@@ -26,6 +26,8 @@
 #
 #
 
+from __future__ import absolute_import
+from __future__ import print_function
 from qscripts_config import __version__, QScriptsConfig as QScfg
 
 import sys
@@ -99,7 +101,7 @@ args = parser.parse_args()
 if args.cont == None and args.restraint == None:
     args.restraint = 'top'
 elif args.cont != None and args.restraint == None:
-    print "FATAL! --rest is required with --cont"
+    print("FATAL! --rest is required with --cont")
     sys.exit(1)
 
 kwargs = {"relax_proc_file": args.relax_proc,
@@ -116,7 +118,7 @@ try:
     gen_inps = genrelax(**kwargs)
     #print gen_inps
 except QGenrelaxError as err:
-    print "ERROR: {}".format(err)
+    print("ERROR: {}".format(err))
     sys.exit(1)
 
 

[…]
