### 2018-09-26 5b9c539
qscripts-cli: tests work with py27 and py37
