### 2018-09-26 a21a0d2
Added the 'six' package, updated travis conf.
