+++ .travis.yml
@@ -1,6 +1,10 @@
 language: python
 python:
   - "2.7"
+  - "3.4"
+  - "3.5"
+  - "3.6"
+  - "3.7"
 before_install:
   - pip install pytest pytest-cov
   - pip install coveralls
