### 2018-09-26 d10a94f
Works locally on 2.7.6, 3.4.9, 3.7.0
