+++ qscripts-cli/q_makefep.py
@@ -24,8 +24,10 @@
 # SOFTWARE.
 #
 
-from __future__ import absolute_import
-from __future__ import print_function
+from __future__ import absolute_import, print_function
+from __future__ import division, unicode_literals
+import six
+
 from qscripts_config import __version__, QScriptsConfig as QScfg
 
 import sys
@@ -35,7 +37,6 @@ import logging
 
 from Qpyl.qmakefep import make_fep, QMakeFepError
 from Qpyl.common import backup_file, init_logger, get_version_full
-import six
 
 
 if __name__ == "__main__":
