+++ qscripts-cli/q_genrelax.py
@@ -26,8 +26,9 @@
 #
 #
 
-from __future__ import absolute_import
-from __future__ import print_function
+from __future__ import absolute_import, print_function
+from __future__ import division, unicode_literals
+
 from qscripts_config import __version__, QScriptsConfig as QScfg
 
 import sys
