+++ qscripts-cli/q_calc.py
@@ -24,10 +24,10 @@
 # SOFTWARE.
 #
 
-from __future__ import unicode_literals
-from __future__ import absolute_import
-from __future__ import print_function
+from __future__ import absolute_import, print_function
+from __future__ import division, unicode_literals
 from io import open
+import six
 
 from qscripts_config import __version__, QScriptsConfig as QScfg
 
@@ -39,7 +39,6 @@ import argparse
 from Qpyl.qgroupcontrib import QGroupContrib, QGroupContribError
 from Qpyl import plotdata
 from Qpyl.common import backup_file, init_logger, get_version_full
-import six
 
 def gc(args):
 
