+++ qscripts-cli/q_analysedyns.py
@@ -28,8 +28,8 @@
 # Extracts dynamics information from the logfile and saves to json
 #
 
-from __future__ import absolute_import
-from __future__ import print_function
+from __future__ import absolute_import, print_function
+from __future__ import division, unicode_literals
 from qscripts_config import __version__, QScriptsConfig as QScfg
 
 import sys
