### 2018-09-26 d9b5d96
Fixed issues with configparser (py27 ignores unicode, py3x ignores inline comments)
