### 2018-09-27 18860f2
Updated travis.yml to include only existing python versions.
