+++ .travis.yml
@@ -3,8 +3,10 @@ python:
   - "2.7"
   - "3.4"
   - "3.5"
+  - "3.5-dev"
   - "3.6"
-  - "3.7"
+  - "3.6-dev"
+  - "3.7-dev"
 before_install:
   - pip install pytest pytest-cov
   - pip install coveralls
