### 2018-09-27 7f88932
Local Q binaries can be used by setting the QBIN_DIR variable.
