### 2018-09-27 baf501d
Fixed the tests for OSX, py2.7
