### 2018-10-01 3b939e6
README cleanup
