### 2018-10-02 9778a16
Renamed 'np' to 'stats', fixed the tests, cleaned up Datacontainer.
The stats class provides functions for calculating the mean, median,
stdev (sample) and sem (st. error of mean).
In Python3+, the first three are simply wrapping the functions in the
statistics library, returning float('nan') on too few data points
instead of raising StatisticsError.
