+++ packages/Qpyl/common.py
@@ -39,6 +39,11 @@ import shutil
 import logging
 import gzip
 
+try:
+    import statistics
+except ImportError:
+    pass
+
 logger = logging.getLogger(__name__)
 
 from Qpyl import __version__
@@ -152,9 +157,8 @@ def backup_file(filename):
     return ""
     
 
-
 # no need for numpy to do these basic stats
-class np():
+class stats(object):
 
     @staticmethod
     def mean(vals):
@@ -162,40 +166,53 @@ class np():
 
         Args:
             vals (list of float):  sample values
+
+        Wraps statistics.mean() in Py3+.
+
+        Returns float('nan') on empty array.
         """
-        N = len(vals)
-        if N == 0: 
+        if len(vals) == 0:
             return float('nan')
-        return sum(vals) * 1.0 / N
+        try:
+            return statistics.mean(vals)
+        except NameError:
+            return 1.0 / len(vals) * sum(vals)
+
 
     @staticmethod
-    def std(vals, ddof=1):
-        """Calculate standard deviation.
+    def stdev(vals):
+        """Calculate sample standard deviation.
         
         Args:
             vals (list of float):  sample values
-            ddof (int, optional):  delta degrees of freedom (default is 1, \
-                                   producing *sample* standard deviation)
+
+        Wraps statistics.stdev() in Py3+.
+
+        Returns float('nan') when fewer than two values.
         """
-        N = len(vals)
-        if N == 0 or N-ddof == 0: 
+        if len(vals) < 2:
             return float('nan')
-        mean = np.mean(vals)
-        variance = [(x - mean)**2 for x in vals]
-        return math.sqrt(sum(variance)/(N - ddof))
+
+        try:
+            return statistics.stdev(vals)
+        except NameError:
+            mean = stats.mean(vals)
+            variance = [(x - mean)**2 for x in vals]
+            return math.sqrt(sum(variance)*1.0/(len(vals)-1))
 
     @staticmethod
-    def std_error(vals, ddof=1):
+    def sem(vals):
         """Calculates standard error of mean.
         
         Args:
             vals (list of float):  sample values
-            ddof (int, optional):  see np.std()
+
+        Returns float('nan') when fewer than two values.
         """
-        N = len(vals)
-        if N == 0 or N-ddof == 0:
+        if len(vals) < 2:
             return float('nan')
-        return np.std(vals, ddof=ddof) / math.sqrt(N)
+
+        return stats.stdev(vals) / math.sqrt(len(vals))
 
     @staticmethod
     def median(vals):
@@ -203,16 +220,22 @@ class np():
 
         Args:
             vals (list of float):  sample values
+
+        Wraps statistics.median() in Py3+.
+
+        Returns float('nan') on empty array.
         """
         N = len(vals)
         if N == 0:
             return float('nan')
-        vals = sorted(vals)
-        if N % 2 == 0: #even
-            return np.mean((vals[N//2-1], vals[N//2]))
-        else: #odd
-            return vals[N//2]
-
+        try:
+            return statistics.median(vals)
+        except NameError:
+            vals = sorted(vals)
+            if N % 2 == 0: #even
+                return stats.mean((vals[N//2-1], vals[N//2]))
+            else: #odd
+                return vals[N//2]
 
 
 
@@ -231,14 +254,30 @@ class DataContainer(object):
         coltitles (list): column titles
 
     Examples:
-        >>> dg_de = DataContainer( ['Energy_gap', 'dG'] )
-        >>> dg_de.add_row( [-300.0, 10.0 ]
[…]
