### 2018-10-02 #935 @fdroessler
Sorry I wanted to check if the following item: ` Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?` refers to the references in the paper or the reference of the software itself.
