### 2018-10-02 #935 @fdroessler
@kyleniemeyer @mpurg I have been through my checklist now and will wait till some of the issues that @malramsay64 and myself have raised are addressed.
