### 2018-10-10 #935 @malramsay64
@mpurg The updates you have been making to the code are all great. There is still the documentation which needs more explanation aimed at a new user. Feel free to get in contact if you want some ideas or feedback as you make changes.
