### 2018-10-10 #935 @mpurg
Hi @malramsay64, @kyleniemeyer, @fdroessler
I've already addressed some of the issues as well as ported the whole codebase to Py3+, but as @malramsay64 points out, the documentation needs quite some work. Unfortunately, I'm juggling this revision with studying for a final exam and preparing for my PhD defense, making the whole process at this moment painfully slow and inefficient. Considering the extent of the revisions, would it be possible to postpone the revision process until the 3rd of Nov?
