### 2018-11-05 #935 @kyleniemeyer
Hi @mpurg, just wanted to check on the status of your revisions—hope your PhD defense went well!
