### 2018-11-06 #935 @mpurg
It went well, thanks.
I'll go over the revisions in the following days, I should be done in a
week or so.
