### 2019-03-29 f4b5698
Fix for q_amber2q.py script.
Torsions and impropers with generic parameters resulted
in multiple identical parameters for the same atom types.

 e.g. for benzene the following

?    ca    ca     ?       14.5   2.0 180.0   4.0

gives

ca    ca    ca     ca       14.5   2.0 180.0   4.0
ca    ca    ca     ca       14.5   2.0 180.0   4.0
ca    ca    ca     ca       14.5   2.0 180.0   4.0
ca    ca    ca     ca       14.5   2.0 180.0   4.0
ca    ca    ca     ca       14.5   2.0 180.0   4.0
ca    ca    ca     ca       14.5   2.0 180.0   4.0

Since identical torsion parameters are not overridden in Q 
[…]
