### 2019-05-20 #935 @mpurg
Hi all, I agree that the submission be withdrawn and apologize for any
inconvenience.
Thank you reviewers for your valuable feedback.

On Mon, May 20, 2019 at 10:08 PM Lorena A. Barba <notifications@github.com>
wrote:

> If this submission is withdrawn (and given the long delay, I agree that it
> should), we would ask the authors to please provide a link to this issue
> when they submit again (via a comment in the Pre-Review issue). This way,
> we can recall the history and possibly ask the same reviewers if they are
> still willing to contribute a review.
>
